ra-raw: make initial receive sync optional
Newer firmwares do not send anything until they receiv at least one
valid frame on UART. So we'd wait forever, so let's make this
optional.
The usual use-case is that the safety chip is resetted, then
wait for the first frame via UART and then starts sending
on its own.
This is triggered on our side by the query of firmware/Git hash.

Signed-off-by: Michael Heimpold <michael.heimpold@chargebyte.com>