Add proper version info support
This adds the cmake magic to include proper project and version
information in the compiled binaries.

For this, a helper include file is generated which is filled with
cmake variables.

In case the sources are built from a git checkout, the git version
string as reported by git itself is used in the PACKAGE_STRING define,
otherwise the cmake project stuff is used as fallback.

Signed-off-by: Michael Heimpold <michael.heimpold@chargebyte.com>